requirements.txt
Django==2.1
djangorestframework
pylint
isort